src/requirements.txt
fastapi[all]
typing
uvicorn
jinja2
python-dateutil